Project_01/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
clerk-sdk-python
openai
clerk_backend_api